iac-analyzer/requirements.txt
python-hcl2==3.0.3